pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gemini-deep-research-mcp"
version = "0.1.0"
description = "MCP server exposing Gemini Deep Research (Interactions API) tools"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Gemini Deep Research MCP" }]
dependencies = [
  "mcp>=1.2.0",
  "google-genai>=0.6.0",
  "python-dotenv>=1.0.1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
]

[project.scripts]
gemini-deep-research-mcp = "gemini_deep_research_mcp.server:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
